+++ adapter/spark/src/main/java/com/asml/apa/wta/spark/driver/WtaDriverPlugin.java
@@ -63,10 +63,13 @@ public class WtaDriverPlugin implements DriverPlugin {
    */
   @Override
   public Map<String, String> init(SparkContext sparkCtx, PluginContext pluginCtx) {
+    log.info("Initialising WTA driver plugin.");
     Map<String, String> executorVars = new HashMap<>();
     try {
-      String configFile =
-          sparkCtx.conf().get("spark.driver.extraJavaOptions").split("-DconfigFile=")[1];
+      String configFile = sparkCtx.conf()
+          .get("spark.driver.extraJavaOptions")
+          .split("-DconfigFile=")[1]
+          .split(" ")[0];
       RuntimeConfig runtimeConfig = RuntimeConfig.readConfig(configFile);
       this.metricStreamingEngine = new MetricStreamingEngine();
       sparkDataSource = new SparkDataSource(sparkCtx, runtimeConfig);
@@ -94,6 +97,7 @@ public class WtaDriverPlugin implements DriverPlugin {
    */
   @Override
   public Object receive(Object message) {
+    log.trace("The driver received a message from an executor.");
     if (message instanceof ResourceCollectionDto) {
       ((ResourceCollectionDto) message)
           .getResourceCollection()
@@ -123,6 +127,7 @@ public class WtaDriverPlugin implements DriverPlugin {
     } catch (Exception e) {
       log.error("A {} error occurred while generating files: {}", e.getClass(), e.getMessage());
     }
+    log.info("WTA driver plugin shutting down.");
   }
 
   /**
@@ -132,6 +137,7 @@ public class WtaDriverPlugin implements DriverPlugin {
    * @since 1.0.0
    */
   private void endApplicationAndWrite() {
+    log.debug("Attempting to write data to file.");
     removeListeners();
     List<Task> tasks = sparkDataSource.getRuntimeConfig().isStageLevel()
         ? sparkDataSource.getStageLevelListener().getProcessedObjects()
@@ -156,6 +162,7 @@ public class WtaDriverPlugin implements DriverPlugin {
     wtaWriter.write(workload);
 
     Stream.deleteAllSerializedFiles();
+    log.debug("Successfully wrote data to file.");
   }
 
   /**
@@ -165,8 +172,12 @@ public class WtaDriverPlugin implements DriverPlugin {
    * @since 1.0.0
    */
   public void initListeners() {
+    log.trace("Initializing listeners.");
     if (!sparkDataSource.getRuntimeConfig().isStageLevel()) {
       this.sparkDataSource.registerTaskListener();
+      log.info("Task level metrics are enabled.");
+    } else {
+      log.info("Stage level metrics are enabled.");
     }
     this.sparkDataSource.registerStageListener();
     this.sparkDataSource.registerJobListener();
@@ -180,9 +191,8 @@ public class WtaDriverPlugin implements DriverPlugin {
    * @since 1.0.0
    */
   public void removeListeners() {
-    if (!sparkDataSource.getRuntimeConfig().isStageLevel()) {
-      sparkDataSource.removeTaskListener();
-    }
+    log.trace("Removing listeners.");
+    sparkDataSource.removeTaskListener();
     sparkDataSource.removeStageListener();
     sparkDataSource.removeJobListener();
     sparkDataSource.removeApplicationListener();