+++ core/src/main/java/com/asml/apa/wta/core/WtaWriter.java
@@ -56,6 +56,7 @@ public class WtaWriter {
    * @since 1.0.0
    */
   public void write(Workload workload) {
+    log.debug("Writing workload to file.");
     try (JsonWriter<Workload> workloadWriter = createWorkloadWriter()) {
       workloadWriter.write(workload);
     } catch (IOException e) {
@@ -73,6 +74,7 @@ public class WtaWriter {
    * @since 1.0.0
    */
   public <T extends BaseTraceObject> void write(Class<T> clazz, List<T> wtaObjects) {
+    log.debug("Writing objects of type {} to file.", clazz.getSimpleName());
     String label = parquetLabels.get(clazz);
     ParquetSchema schema = new ParquetSchema(clazz, wtaObjects, label);
     OutputFile path = file.resolve(label).resolve(schemaVersion).resolve(label + ".parquet");
@@ -93,6 +95,7 @@ public class WtaWriter {
    * @since 1.0.0
    */
   protected void setupDirectories(OutputFile path, String version) {
+    log.debug("Setting up directory structure for the output.");
     try {
       path.resolve("workload").resolve(version).resolve(".temp").clearDirectory();
       for (String directory : parquetLabels.values()) {