+++ core/src/main/java/com/asml/apa/wta/core/streams/Stream.java
@@ -134,6 +134,7 @@ public class Stream<V extends Serializable> {
    * @since 1.0.0
    */
   private synchronized void serializeInternals() {
+    log.trace("Serializing stream {}.", id);
     StreamNode<V> current;
     if (head == deserializationEnd) {
       current = head.getNext();
@@ -150,7 +151,7 @@ public class Stream<V extends Serializable> {
     try (ObjectOutputStream objectOutputStream = new ObjectOutputStream(new FileOutputStream(filePath))) {
       objectOutputStream.writeObject(toSerialize);
     } catch (IOException e) {
-      log.error("Failed to serialize stream internals to {}", filePath);
+      log.error("Failed to serialize stream internals to {}.", filePath);
       return;
     }
     deserializationEnd.setNext(null);
@@ -169,6 +170,7 @@ public class Stream<V extends Serializable> {
    * @since 1.0.0
    */
   private synchronized int deserializeInternals(@NonNull String filePath) {
+    log.trace("Deserializing stream internals from {}.", filePath);
     int amountOfNodes;
     try (ObjectInputStream objectInputStream = new ObjectInputStream(new FileInputStream(filePath))) {
       List<StreamNode<V>> nodes = (ArrayList<StreamNode<V>>) objectInputStream.readObject();
@@ -187,7 +189,7 @@ public class Stream<V extends Serializable> {
       }
       amountOfNodes = nodes.size();
     } catch (IOException | ClassNotFoundException | ClassCastException e) {
-      log.error("Failed to deserialize stream internals from {}", filePath);
+      log.error("Failed to deserialize stream internals from {}.", filePath);
       throw new FailedToDeserializeStreamException();
     } finally {
       new File(filePath).delete();
@@ -203,6 +205,7 @@ public class Stream<V extends Serializable> {
    * @since 1.0.0
    */
   public boolean isEmpty() {
+    log.trace("Checking whether stream {} is empty.", id);
     return head == null;
   }
 
@@ -217,8 +220,9 @@ public class Stream<V extends Serializable> {
    * @since 1.0.0
    */
   public synchronized V head() {
+    log.trace("Head of stream {} was requested.", id);
     if (head == null) {
-      log.error("Stream#head() was called on an empty stream");
+      log.error("`Stream#head()` was called on an empty stream.");
       throw new NoSuchElementException();
     }
     additionsSinceLastWriteToDisk--;
@@ -246,8 +250,9 @@ public class Stream<V extends Serializable> {
    * @since 1.0.0
    */
   public synchronized V peek() {
+    log.trace("Peeked at head of stream {}", this.id);
     if (head == null) {
-      log.error("Stream#peek() was called on an empty stream");
+      log.error("`Stream#peek()` was called on an empty stream.");
       throw new NoSuchElementException();
     }
     return head.getContent();
@@ -261,6 +266,7 @@ public class Stream<V extends Serializable> {
    * @since 1.0.0
    */
   public synchronized void addToStream(V content) {
+    log.trace("Added content to stream {}", this.id);
     if (head == null) {
       head = new StreamNode<>(content);
       tail = head;
@@ -273,7 +279,7 @@ public class Stream<V extends Serializable> {
     additionsSinceLastWriteToDisk++;
     if (additionsSinceLastWriteToDisk > serializationTrigger) {
       log.trace(
-          "Serializing stream internals after {} additions since last write to disk",
+          "Serializing stream internals after {} additions since last write to disk.",
           additionsSinceLastWriteToDisk);
       serializeInternals();
     }
@@ -292,6 +298,7 @@ public class Stream<V extends Serializable> {
    * @since 1.0.0
    */
   public synchronized <R extends Serializable> Stream<R> map(@NonNull Function<V, R> op) {
+    log.trace("Consuming and applying map on stream {}", this.id);
     StreamNode<V> next = head;
     Stream<R> ret = new Stream<>();
     while (next != null) {
@@ -319,6 +326,7 @@ public class Stream<V extends Serializable> {
    * @since 1.0.0
    */
   public synchronized Stream<V> filter(@NonNull Function<V, Boolean> predicate) {
+    log.trace("Consuming and applying filter on stream {}", this.id);
     StreamNode<V> next = head;
     Stream<V> ret = new Stream<>();
     while (next != null) {
@@ -350,6 +358,7 @@ public class Stream<V extends Serializable> {
    * @since 1.0.0
    */
   public synchronized <R> R foldLeft(R init, @NonNull BiFunction<R, V, R> op) {
+    log.trace("Consuming and applying left fold on stream {}", this.id);
     R acc = init;
     StreamNode<V> next = head;
     while (next != null) {
@@ -373,6 +382,7 @@ public class Stream<V extends Serializable> {
    * @since 1.0.0
    */
   public synchronized List<V> toList() {
+    log.trace("Consuming stream {} to list", this.id);
     List<V> ret = new ArrayList<>();
     while (!isEmpty()) {
       ret.add(head());